🚚 [RUM] move trackTimings code to a separate module
