+++ packages/rum/src/domain/rumEventsCollection/view/trackViews.ts
@@ -3,6 +3,7 @@ import { addGlobalEventListeners, DOM_EVENT, generateUUID, monitor, ONE_MINUTE, 
 import { LifeCycle, LifeCycleEventType } from '../../lifeCycle'
 import { EventCounts, trackEventCounts } from '../../trackEventCounts'
 import { waitIdlePageActivity } from '../../trackPageActivities'
+import { Timings, trackTimings } from './trackTimings'
 
 export interface View {
   id: string
@@ -22,14 +23,6 @@ export interface ViewCreatedEvent {
   location: Location
   referrer: string
   startTime: number
-}
-
-export interface Timings {
-  firstContentfulPaint?: number
-  domInteractive?: number
-  domContentLoaded?: number
-  domComplete?: number
-  loadEventEnd?: number
 }
 
 export enum ViewLoadingType {
@@ -212,32 +205,6 @@ function trackHistory(onHistoryChange: () => void) {
 
 function trackHash(onHashChange: () => void) {
   addGlobalEventListeners([DOM_EVENT.HASH_CHANGE], onHashChange)
-}
-
-function trackTimings(lifeCycle: LifeCycle, callback: (timings: Timings) => void) {
-  let timings: Timings | undefined
-  const { unsubscribe: stopPerformanceTracking } = lifeCycle.subscribe(
-    LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED,
-    (entry) => {
-      if (entry.entryType === 'navigation') {
-        timings = {
-          ...timings,
-          domComplete: entry.domComplete,
-          domContentLoaded: entry.domContentLoadedEventEnd,
-          domInteractive: entry.domInteractive,
-          loadEventEnd: entry.loadEventEnd,
-        }
-        callback(timings)
-      } else if (entry.entryType === 'paint' && entry.name === 'first-contentful-paint') {
-        timings = {
-          ...timings,
-          firstContentfulPaint: entry.startTime,
-        }
-        callback(timings)
-      }
-    }
-  )
-  return { stop: stopPerformanceTracking }
 }
 
 function trackLoadingTime(loadType: ViewLoadingType, callback: (loadingTime: number) => void) {
