+++ packages/rum/src/types.ts
@@ -1,7 +1,8 @@
 import { Context, ErrorSource, ResourceType } from '@datadog/browser-core'
 import { ActionCounts, ActionType } from './domain/rumEventsCollection/action/trackActions'
 import { PerformanceResourceDetails } from './domain/rumEventsCollection/resource/resourceUtils'
-import { Timings, ViewLoadingType } from './domain/rumEventsCollection/view/trackViews'
+import { Timings } from './domain/rumEventsCollection/view/trackTimings'
+import { ViewLoadingType } from './domain/rumEventsCollection/view/trackViews'
 import { EventCounts } from './domain/trackEventCounts'
 
 export enum RumEventCategory {
